A generated data set for test

A generated data set provided to test the Interatrix package.

Usage

data(dataInteratrix)

Format

A data frame with 100 observations for the following variables:

F1

a numeric vector containing a factor with three modalities

F2

a numeric vector containing a continuous variable

F3

a numeric vector containing a factor with two modalities

F4

a numeric vector containing a continuous variable

Parasite1

a numeric vector containing the serological status to the first parasite

Parasite2

a numeric vector containing the serological status to the second parasite

AGE

a numeric vector containing a factor with three modalities indicating the age classes

Examples

data(dataInteratrix)